rename abi.js -> proxyContractsABI.js
